Cicd: Don't commit version bumps on CI/CD branches
CI/CD branches are for testing the build system, not incrementing
the version. The minor/fix version in `Cargo.toml` will still
increase, but the change won't live beyond this branch. This
has the potential to create duplicate releases, but since CI/CD
branches only create draft releases, this shouldn't be an issue.